fixed discover crash and revised layout of suggested people

+++ victorious/victorious/UI/Discover/Cells/VSuggestedPeopleCell.h
@@ -10,6 +10,6 @@
 
 @interface VSuggestedPeopleCell : UITableViewCell
 
-+ (NSInteger)cellHeight;
++ (CGFloat)cellHeight;
 
 @end
